requirements.txt
astropy==6.1.0
jax==0.4.29
jaxlib==0.4.29
matplotlib==3.8.4
numpy==1.26.4
photutils==1.12.0
pyregion==2.2.0
PyYAML==6.0.1
regions==0.9
reproject==0.13.1
scipy==1.13.0
astroquery==0.4.7
pydantic==2.10.4
